Migrating HTML-driven banners to Hugo
We originally built the images with autobot. However, since the website
already has the necessary metadata and Hugo supports it (with Hugo
Custom Content Types) – it makes sense to migrate this generation to
Hugo. We have to screenshot the pages anyways (they were originally
HTML-based), this just makes it easier for other things like the OBS
system to retrieve up-to-date HTML pages.