Figured out auto-generation of Group profiles
Needed to iron out how information is laid out on the website. Finished
off adding shortcodes for displaying the who/what/when/where of the
Group's meeting. Also added a proper list of the syllabus. Finished off
with a list of the semester's coordinators.

Originally tried using HTML templates, but this failed as the goal was
to also display a ToC of the page (since it gets pretty long). Shifted
back to displaying it from the <group>/<semester>/_index.md file
instead.